Survey Clean up Work Part 1 - Refresh Survey List for Create, Edit, Delete functions
This code ensures that the list of surveys is refreshed (the page does
NOT refresh) whenever the CREATE, DELETE or UPDATE button is
used.
Also cleaned up the table layout in survey handlebars and
removed the ID field.

Next up:
Take ID out outUpdate modal.
Hide Responses whenever a survey is deleted.